device/DUT/requirements.txt
# --- Webserver & API ---
flask
apiflask
requests
pyyaml

# --- Database (psycopg 3) ---
psycopg[binary]

# --- Raspberry Pi hardware libraries ---
gpiozero
smbus2
pi-ina219
luma.oled
rpi_ws281x